Users should work. And you can make people admins
Link to users added in the bottom of forum.php
Made the space for user, firstname, lastname and email four times as big as
before, to make space for plaucible < or >, which is now replaced with
'&lt;' or '&gt;' so you cant make tags in your names and such... Should this
feature be deleted entirely, and instead ban having any < or > in the
names?
For now to variables with string arrays of what should be replaced with what
is placed in the files where they are used. Perhaps they should be in util
instead?
The choice of replacing the tags in the table, is so that we don't have to
remeber it later on, when we print the names.